ActionType choose_speed (again) (#39)
* Reworked controllers to ActionController

All methods are currently in ActionController. No controllers have been deleted yet.

* Various bug fixes

Oh boy it's a swarm of bugs out here. Run for your lives!

* deleted print statement

Accidentally left a print statement in engine.py

* Deleted ActionType change_speed

change_speed is already a function in Truck and doesn't need to be an action_type since it's just changing a value on the truck. Therefore, it's not necessary and is deleted from the ActionType enum and from the Action_Controller

* ActionType choose_speed (again)

My previous PR said this wasn't needed since the Truck had the set_current_speed method. However, this will not work because the user client is given a copy of the truck and cannot directly change its values.